Reset visualizations when node is reset (#842)
There was a problem that sometimes placeholder hasn't change despite on node being gray out (which indicates that we know nothing and we wait for info from backend). This also fix a problem where too fast clicking on Next button in tutorial causes unwanted visualization toggle - it was happening because of delayed responses from backend that caries outdated informations.