Revert "install manager with luna-studio (#180)"
This reverts commit 472ce6434023a60ce65d03066a9b433596865f3a.
We've decided to revert it, because it was a part of the effort
to install all Luna-related apps with Luna Manager. The effort,
however, was never quite pushed to any conclusion and the changes
here are very backwards incompatible, requiring us to change all
of the configs and breaking existing installs.